requirements.txt
Flask==2.0.3
Flask-Mail==0.9.1
python-dotenv==0.19.2
